ml-service/requirements.txt
pandas
numpy
scikit-learn
flask
python-dotenv
celery
redis